+++ src/presentation/middlewares/auth-middleware-protocols.ts
@@ -1,0 +1,3 @@
+export * from '../protocols'
+export * from '../../domain/usecases/load-account-by-token'
+export * from '../../domain/models/account'
