+++ src/presentation/protocols/index.ts
@@ -1,3 +1,4 @@
 export * from './controller'
 export * from './http'
 export * from './validation'
+export * from './middleware'
